chore(wizard-refactor): extract tools/update_course_config
Post-creation educator tuning path. No pinning tests; dispatcher 28/28 green.
